requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-huggingface
chromadb
transformers
sentence-transformers
pypdf
loguru
python-dotenv
streamlit
httpx
requests
pandas
pydantic
typing-extensions
openai
together
langchain_chroma
langchain_together
langchain_community